Add idemponent resource group command
The group command makes it simpler to ensure a resource group
exists, and that it has been cleaned up when the bundle is uninstalled.

When used in any action other than uninstall, the mixin will ensure that the resource group exists.
```yaml
install:
  - az:
      description: "Ensure my group exists"
      group:
        name: mygroup
        location: westus
```

When used in the uninstall action, the mixin ensures that the resource group is deleted.
```yaml
uninstall:
  - az:
      description: "Cleanup my group"
      group:
        name: mygroup
```

I have also added the ignoreError functionality from the exec mixin so
that custom user commands can handle errors.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>