Change installed az CLI version with clientVersion
I've added clientVersion to the az mixin configuration so that you can specify which version of the az CLI to install when porter builds the bundle. By default, the latest version of the CLI is installed.

    ```yaml
    mixins:
    - az:
        clientVersion: 2.35.0
    ```

The mixin will use the clientVersion to build the apt package version published by Microsoft, which follows the format VERSION-1~DISTROCODENAME. So in the example above, the package version is 2.35.0-1~stretch. If the user changes the debian version, then stretch will change to match the version, e.g. buster, etc.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>