Set azure user agent environment variable
When the bundle is built, the az mixin sets the azure cli user agent environment variable AZURE_HTTP_USER_AGENT in the Dockerfile so that the user agent will be passed either when the mixin calls the az cli directly, or when the bundle uses the az cli in a bash helper script.

This avoids gaps where some of the bundle uses the mixin and other complex parts use a script, which can make it more difficult to debug when we don't get requests from both flagged as coming from porter.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>